requirements.txt
fastapi==0.112.0
openai==1.40.3
pydantic==2.8.2
pymongo==4.8.0
python-dotenv==1.0.1
uvicorn==0.30.5
